+++ sdk/CHANGELOG.md
@@ -23,9 +23,12 @@ This changelog summarizes major changes between GraalVM SDK versions. The main f
     * Added `getHashSize()` to return hash entries count.
     * Added `hasHashEntry(Object)` specifying that the mapping for the specified key exists.
     * Added `getHashValue(Object)` returning the value for the specified key.
+    * Added `getHashValueOrDefault(Object, Object)` returning the value for the specified key or a default value if the mapping for given key does not exist.
     * Added `putHashEntry(Object, Object)` associating the specified value with the specified key.
     * Added `removeHashEntry(Object)` removing the mapping for a given key.
     * Added `getHashEntriesIterator()` returning a hash entries iterator.
+    * Added `getHashKeysIterator()` returning a hash keys iterator.
+    * Added `getHashValuesIterator()` returning a hash values iterator.
 * Added `HostAccess.Builder.allowMapAccess(boolean)` to allow the guest application to access Java `Map` as values with hash entries (true by default for `HostAccess.ALL`, false otherwise).
 * Added `ProxyHashMap` to proxy map guest values.
 * When `HostAccess.Builder.allowMapAccess(boolean)` is enabled the Java `HashMap.Entry` is interpreted as a guest value with with two array elements.